refactor: improve logger construction
