+++ internal/httpserver/config.go
@@ -20,16 +20,12 @@ type Server struct {
 	log     *slog.Logger
 }
 
-func NewService(conf Config) service.Service {
+func NewService(conf Config, logger *slog.Logger) service.Service {
 	name := "HTTP"
 	if conf.TLS != nil {
 		name = "HTTPS"
 	}
-	return &Server{name: name, conf: conf}
-}
-
-func (s *Server) SetLogger(logger *slog.Logger) {
-	s.log = logger
+	return &Server{name: name, conf: conf, log: service.NewPrefixedLogger(logger, name)}
 }
 
 type Config struct {
