+++ internal/service/service.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"context"
-	"log/slog"
 )
 
 type Service interface {
@@ -14,10 +13,4 @@ type Service interface {
 
 	// stop gracefully shuts down the service within the given context deadline
 	Stop(ctx context.Context) error
-}
-
-// optional interface implemented by services that want
-// a logger injected by the service manager
-type LoggerAware interface {
-	SetLogger(logger *slog.Logger)
 }
